Fixes and LatencyBench
This patch fixes ZoomStorageTest and introduces a new Benchmark called
LatencyBench. This benchmark initializes a map/set of a given size and
has one client issue a number of writes to it. The benchmark reports the
average latency between the write and the time when we receive the
updated map/set.

This is a preliminary version of LatencyBench.