Children deletion fix
The benchmarks always try to clean any left over children from previous
executions. Since children are epheral nodes, it could happen that
between the time when we obtian the list of children and when we delete
them, some have been deleted. We now take into account such cases and
simply ignore exceptions raised because a child does not exist.