Bug fixes in collectResults & logging
This patch fixes some bugs in collectResults where indices to retrieve
the updater and client timestamps for the data structure versions were
incorrect. I also now log the timestamps of updates and reception of the
data structure, as well as the timestamps collected by the mpi root in
method collectResults.