Computing the average latency for ChurnBench differently
This patch makes the computation of the average latency for ChurnBench
more "fair". Now the latency is simply the time difference between the
update and the reception of the table by any client. Previously, the
latency was the time difference between the update and the reception of
the table by all clients.